# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Traveluxe OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: client_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a new quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a quote by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
    put:
      operationId: updateQuote
      tags: [quotes]
      summary: Update a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}/convert:
    post:
      operationId: convertQuoteToBooking
      tags: [quotes]
      summary: Convert a quote to a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created booking from quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List all drivers
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create a new driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverBody"
      responses:
        "201":
          description: Created driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get a driver by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Driver details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverDetail"
    put:
      operationId: updateDriver
      tags: [drivers]
      summary: Update a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverBody"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}/rate:
    post:
      operationId: rateDriver
      tags: [drivers]
      summary: Rate a driver after a completed job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateDriverBody"
      responses:
        "201":
          description: Rating created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverRating"
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a channel or DM
      parameters:
        - name: channel
          in: query
          schema:
            type: string
        - name: recipient_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/complete:
    put:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    Quote:
      type: object
      properties:
        id:
          type: string
        client_id:
          type: string
        client_name:
          type: string
        service_type:
          type: string
        direction:
          type: string
        pickup:
          type: string
        dropoff:
          type: string
        date_time:
          type: string
        passengers:
          type: integer
        vehicle_type:
          type: string
        duration:
          type: number
        destination:
          type: string
        price:
          type: number
        status:
          type: string
        notes:
          type: string
        created_at:
          type: string
        created_by:
          type: string
      required: [id, service_type, price, status]
    CreateQuoteBody:
      type: object
      properties:
        client_id:
          type: string
        client_name:
          type: string
        service_type:
          type: string
        direction:
          type: string
        pickup:
          type: string
        dropoff:
          type: string
        date_time:
          type: string
        passengers:
          type: integer
        vehicle_type:
          type: string
        duration:
          type: number
        destination:
          type: string
        price:
          type: number
        status:
          type: string
        notes:
          type: string
      required: [service_type, price]
    Booking:
      type: object
      properties:
        id:
          type: string
        tvl_ref:
          type: string
        client_id:
          type: string
        client_name:
          type: string
        client_vip_tier:
          type: string
        quote_id:
          type: string
        service_type:
          type: string
        direction:
          type: string
        pickup:
          type: string
        dropoff:
          type: string
        destination:
          type: string
        flight_number:
          type: string
        date_time:
          type: string
        passengers:
          type: integer
        luggage:
          type: integer
        vehicle_type:
          type: string
        nameboard:
          type: string
        special_requests:
          type: string
        additional_charges:
          type: number
        price:
          type: number
        tvl_commission:
          type: number
        driver_receives:
          type: number
        commission_type:
          type: string
        payment_status:
          type: string
        payment_method:
          type: string
        commission_status:
          type: string
        payout_status:
          type: string
        source:
          type: string
        status:
          type: string
        operator_id:
          type: string
        operator_name:
          type: string
        driver_id:
          type: string
        driver_name:
          type: string
        driver_vehicle:
          type: string
        return_booking_id:
          type: string
        is_amended:
          type: boolean
        notes:
          type: string
        cancellation_reason:
          type: string
        cancellation_fee:
          type: number
        created_at:
          type: string
        updated_at:
          type: string
        created_by:
          type: string
        duration:
          type: number
      required: [id, tvl_ref, service_type, status, price]
...
    Driver:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        whatsapp:
          type: string
        vehicle_type:
          type: string
        vehicle_model:
          type: string
        plate:
          type: string
        notes:
          type: string
        status:
          type: string
        avg_rating:
          type: number
        total_jobs:
          type: integer
        created_at:
          type: string
      required: [id, name, vehicle_type, status]
    DriverDetail:
      allOf:
        - $ref: "#/components/schemas/Driver"
        - type: object
          properties:
            bookings:
              type: array
              items:
                $ref: "#/components/schemas/Booking"
            ratings:
              type: array
              items:
                $ref: "#/components/schemas/DriverRating"
            commission_ledger:
              type: array
              items:
                $ref: "#/components/schemas/CommissionEntry"
    CreateDriverBody:
      type: object
      properties:
        name:
          type: string
        whatsapp:
          type: string
        vehicle_type:
          type: string
        vehicle_model:
          type: string
        plate:
          type: string
        notes:
          type: string
        status:
          type: string
      required: [name, whatsapp, vehicle_type]
    DriverRating:
      type: object
      properties:
        id:
          type: string
        driver_id:
          type: string
        booking_id:
          type: string
        rating:
          type: integer
        note:
          type: string
        rated_by:
          type: string
        rated_at:
          type: string
      required: [id, driver_id, booking_id, rating]
    RateDriverBody:
      type: object
      properties:
        booking_id:
          type: string
        rating:
          type: integer
        note:
          type: string
      required: [booking_id, rating]
...
    CommissionEntry:
      type: object
      properties:
        booking_id:
          type: string
        tvl_ref:
          type: string
        date:
          type: string
        client_name:
          type: string
        total_fare:
          type: number
        tvl_commission:
          type: number
        driver_receives:
          type: number
        payment_method:
          type: string
        commission_status:
          type: string
        payout_status:
          type: string
      required: [booking_id, tvl_ref, total_fare, tvl_commission, driver_receives]
...
    Message:
      type: object
      properties:
        id:
          type: string
        channel:
          type: string
        sender_id:
          type: string
        sender_name:
          type: string
        recipient_id:
          type: string
        content:
          type: string
        read:
          type: boolean
        created_at:
          type: string
      required: [id, content, created_at]
    SendMessageBody:
      type: object
      properties:
        channel:
          type: string
        recipient_id:
          type: string
        content:
          type: string
      required: [content]
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        assigned_to:
          type: string
        assigned_to_name:
          type: string
        due_date:
          type: string
        priority:
          type: string
        completed:
          type: boolean
        created_by:
          type: string
        created_at:
          type: string
      required: [id, title, priority, completed]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        assigned_to:
          type: string
        due_date:
          type: string
        priority:
          type: string
      required: [title, priority]
...