# lib/api-zod/src/generated/types/createDriverBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveluxe OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDriverBody {
  name: string;
  whatsapp: string;
  vehicle_type: string;
  vehicle_model?: string;
  plate?: string;
  notes?: string;
  status?: string;
}